radar update

+++ SW/src/rover_radar.cpp
@@ -43,7 +43,7 @@ void RoverRadar::sendRaderInfo(float distance, int angle){
   Serial.print(distance);Serial.print(",");
   Serial.println(angle);
 }
-float RoverRadar::turn_angle(int angle){
+int RoverRadar::turn_angle(int angle){
     if(c){
     i=i+10;
     if(i>160)c=false;
